requirements.txt
streamlit
openai-whisper
google-generativeai
reportlab
speechrecognition
pydub
PyPDF2
soundfile
